+++ index.es6.js
@@ -2,8 +2,13 @@ export { default as default } from './source/input'
 
 export
 {
-	formats as phone_number_format,
-	formats as phoneNumberFormat,
+	default as phone_number_format,
+	default as phoneNumberFormat
+}
+from './source/formats'
+
+export
+{
 	validate as is_valid_phone_number,
 	validate as isValidPhoneNumber,
 	format as format_phone_number,
@@ -16,3 +21,11 @@ export
 	parse_plaintext_international as parsePhoneNumber
 }
 from './source/phone'
+
+export
+{
+	default as country,
+	country_from_locale,
+	country_from_locale as countryFromLocale
+}
+from './source/country'