+++ test/runner.js
@@ -1,7 +1,0 @@
-import chai from 'chai'
-chai.should()
-
-require('./exports')
-
-require('./phone')
-require('./editor')